+++ sillo/mail/__init__.py
@@ -1,5 +1,6 @@
-from .client import MailClient, get_mail_client, setup_mail
+from .client import MailClient, get_mail_client, send_email, setup_mail
 from .config import MailConfig
+from .context import NotConfiguredError, current_mail
 from .models import EmailAttachment, EmailMessage, EmailResult
 
 __all__ = [
@@ -8,6 +9,9 @@ __all__ = [
     "EmailResult",
     "MailClient",
     "MailConfig",
+    "NotConfiguredError",
+    "current_mail",
     "get_mail_client",
+    "send_email",
     "setup_mail",
 ]